+++ changelog.md
@@ -6,3 +6,4 @@
 - On the fly module creation
 - Refrence Schema's via http and inline refrences
 - Support alternatives - oneOf, anyOf
+- Support for dependencies - array of conditionals